convert NumberInput to Ts
